requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
